services/vera-core/pyproject.toml
[build-system]
requires = ["uv>=0.1.0"]
build-backend = "uv.buildapi"

[project]
name = "vera-core"
version = "0.1.0"
description = "Core semantic processing engine for Vera AI Platform"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Vera AI Team"},
]
dependencies = [
    # Основные зависимости
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.21.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.0.0",
    "pydantic[email]>=1.10.7",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "uv>=0.1.0",
    "orjson>=3.8.0",        # Для быстрой работы с JSON
    
    # Базовые ML зависимости
    "scipy>=1.10.0",        # Для работы с разреженными матрицами
    "joblib>=1.2.0",        # Для параллельных вычислений
    "tqdm>=4.65.0",         # Для отображения прогресса
]

[project.optional-dependencies]
# Для разработки
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pylint>=3.0.0",
    "ipython>=8.0.0",
    "pre-commit>=3.0.0",
]

# Для распределенных вычислений
distributed = [
    "dask[complete]>=2023.1.0",
    "distributed>=2023.1.0",
]

# Для ML и анализа данных
ml = [
    "scikit-learn>=1.2.0",
    "pandas>=2.0.0",
    "umap-learn>=0.5.3",
]

# Альтернативный бэкенд для распределенных вычислений
ray = [
    "ray[default]>=2.9.0",
]

# Визуализация
viz = [
    "plotly>=5.13.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

# Интеграция с MCP GitHub
mcp_github = [
    "PyGithub>=1.58.0",
    "python-gitlab>=3.14.0",
    "requests>=2.28.0",
    "cryptography>=39.0.0",
]

[project.scripts]
vera-core = "app.main:main"

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
show_error_context = true
pretty = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "-v --cov=app --cov-report=term-missing"
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[metadata]
# GitHub Actions
requires-gha = ["ubuntu-latest", "windows-latest", "macos-latest"]
python-versions = ["3.10", "3.11"]
requires-dist = ["dask[complete] (>=2023.1.0)", "distributed (>=2023.1.0)"]
requires-python = ">=3.10"
